moved flying ship functionality from spells.c and shipcurse.c to flyingship.c
