+++ src/bind_monsters.c
@@ -1,6 +1,7 @@
 #include <platform.h>
 #include "spells/shipcurse.h"
 #include "monster.h"
+#include "flyingship.h"
 
 #include <kernel/equipment.h>
 #include <kernel/faction.h>
